+++ src/Plugin.php
@@ -16,7 +16,11 @@ use benf\neo\services\Fields as FieldsService;
 use Craft;
 use craft\base\Model;
 use craft\base\Plugin as BasePlugin;
+use craft\console\Controller;
+use craft\console\controllers\ResaveController;
+use craft\db\Query;
 use craft\db\Table;
+use craft\events\DefineConsoleActionsEvent;
 use craft\events\DefineFieldLayoutElementsEvent;
 use craft\events\RebuildConfigEvent;
 use craft\events\RegisterComponentTypesEvent;
@@ -90,6 +94,7 @@ class Plugin extends BasePlugin
         $this->_registerProjectConfigRebuild();
         $this->_registerGarbageCollection();
         $this->_registerChildBlocksUiElement();
+        $this->_registerResaveBlocksCommand();
         $this->_registerGatsbyHelper();
     }
 
@@ -185,7 +190,32 @@ class Plugin extends BasePlugin
         });
     }
 
-    private function _registerGatsbyHelper(): void
+    private function _registerResaveBlocksCommand(): void
+    {
+        Event::on(ResaveController::class, Controller::EVENT_DEFINE_ACTIONS, function(DefineConsoleActionsEvent $event) {
+            $event->actions['neo-blocks'] = [
+                'helpSummary' => 'Re-saves Neo blocks.',
+                'options' => ['field', 'type'],
+                'optionsHelp' => [
+                    'field' => 'The field handle to save Neo blocks for.',
+                    'type' => 'The block type handle(s) of the Neo blocks to resave.',
+                ],
+                'action' => function(): int {
+                    $controller = Craft::$app->controller;
+                    $criteria = [];
+                    if ($controller->field !== null) {
+                        $criteria['field'] = explode(',', $controller->field);
+                    }
+                    if ($controller->type !== null) {
+                        $criteria['type'] = explode(',', $controller->type);
+                    }
+                    return $controller->resaveElements(Block::class, $criteria);
+                }
+            ];
+        });
+    }
+
+    private function _registerGatsbyHelper()
     {
         if (class_exists(Deltas::class)) {
             Event::on(Deltas::class, Deltas::EVENT_REGISTER_IGNORED_TYPES, function(RegisterIgnoredTypesEvent $event) {