+++ src/elements/db/BlockQuery.php
@@ -4,7 +4,6 @@ namespace benf\neo\elements\db;
 
 use benf\neo\elements\Block;
 use benf\neo\models\BlockType;
-use benf\neo\Plugin as Neo;
 use Craft;
 use craft\base\ElementInterface;
 use craft\db\Query;
@@ -89,6 +88,7 @@ class BlockQuery extends ElementQuery
     public function __set($name, $value)
     {
         match ($name) {
+            'field' => $this->field($value),
             'owner' => $this->owner($value),
             'primaryOwner' => $this->primaryOwner($value),
             'type' => $this->type($value),
@@ -103,6 +103,25 @@ class BlockQuery extends ElementQuery
     {
         $this->withStructure = true;
         parent::init();
+    }
+
+    /**
+     * Filters the query results based on the field.
+     *
+     * @param string[]|Field|null $value An array of Neo field handles, a Neo field instance, or `null`
+     * @return $this
+     * @since 2.13.5
+     */
+    public function field($value)
+    {
+        $this->fieldId = !$value ? null : ($value instanceof Field ? [$value->id] : (new Query())
+            ->select(['id'])
+            ->from([Table::FIELDS])
+            ->where(Db::parseParam('handle', $value))
+            ->andWhere(['type' => Field::class])
+            ->column());
+
+        return $this;
     }
 
     /**