+++ CHANGELOG.md
@@ -43,6 +43,12 @@
 ### Fixed
 - Fixed a bug where it was possible to create a Neo field with no block types
 
+## 2.13.5 - 2022-04-16
+
+### Added
+- Added the `php craft resave/neo-blocks` console command
+- Added `benf\neo\elements\db\BlockQuery::field()`
+
 ## 2.13.4 - 2022-04-13
 
 ### Fixed